requirements.txt
Flask==2.0.1
jieba==0.42.1
matplotlib==2.2.5
numpy==1.21.1
Pillow==8.3.1
pymongo==3.12.0
redis==3.5.3
SQLAlchemy==1.4.22
Werkzeug==2.0.1
kafka-python==2.0.2
